move indexes in database package to separate file
